requirements.txt
djangorestframework-jwt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.2
django-polls==0.1
djangorestframework==3.9.4
idna==2.8
isort==4.3.20
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pipenv==2018.11.26
pycodestyle==2.5.0
Pygments==2.4.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
typing==3.6.6
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
wrapt==1.11.1
